dtrace: SDT support on arm64
SDT probes on arm64 demonstrated a flaw in the design of the probe
location collection code.  The argument to the .align directive in
assembler files differs between architectures.  This resulted in
SDT probe data being laid out differently from what the implementation
expected.  As a result, corrupt data was being processed.

The manner in which SDT probe data is stored in the kernel is being
modified with this commit.  Rather than storing each probe as a struct
that contains 3 longs and 2 strings (aligning each struct on a 64-bit
boundary), probe data is stored as an array of 64-bit values (probe
address, probe name offset, and function name offset).  The offsets are
into a string table: a concatenation of all probe and function names.
An added advantage is that this allows for de-duplication of strings.

This change affects all architectures.

In addition, to further implement SDT probes on arm64, a new ESR code
is being added for software breakpoint traps.  This allows us to
associate specific breakpoints with DTrace SDT probe handling code.
The other two existing codes are for kprobes and uprobes.

The logic for the nosdt kernel option changed a little due to the arm64
SDT implementation because it is no longer possible to change the value
of dtrace_sdt_nprobes because it is a constant in RO data.  Instead,
kernel option will now toggle a flag.

Finally, to be able to support a more arch-independent way of collecting
SDT probe points, '--emit-relocs' is now being passed in the temporary
kernel image linking rather than '-r'.  This is in retrospect also more
correct because it is the relocation records we're insterested in.

Due to missing support for > 64k sections when --emit-relocs is used
(see Orabug 28182242), a workaround is included in this commit by means
of the dtrace_sdt_arm64.sh script.  It implements a (slower) alternative
algorithm to generate the SDT probe information that is to be linked into
the kernel.  It is functionally the same, and once the fix for the linker
is available in the toolchain used for kernel builds, this workaround
script can be removed and link_vmlinux.sh can be modified to use the
arch-independent dtrace_sdt.sh script for arm64 as well.

Orabug: 28070261
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>
Reviewed-by: Tom Saeger <tom.saeger@oracle.com>