ctf: avoid command line length limits
Currently, we process the list of object files to manipulate via a huge
shell for loop.  This can overflow after a few thousand object files,
halting the "make ctf".

Fix this by generating a "raw" list of object-files-and-thin-archives
and then using that: we generate the raw list using a make implementation
of xargs using $(wordlist) and $(words).  Since this now does everything
internal to make and via pipelines to "while read" loops, it eliminates
the command-line length problems.

This also fixes the case where the filelist is entirely empty (though
this is unlikely ever to happen).

(Derived from a much older patch from 2018.  This has changed, and it was
ages ago, so Reviewed-by dropped: will need re-review.)

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>