dtrace: FBT entry probes for arm64
This commit provides an implementation of DTrace FBT probes for arm64.
The actual probe support is currently limited to FBT entry probes, as a
first phase.  Probes are fired using a breakpoint that is tagged as a
DTrace FBT breakpoint (similar to DTrace SDT probes).  To this end, the
breakpoint use for DTrace is being modified in this commit also, to be
able to distinguish between SDT and FBT related breakpoints.

Because FBT ends up sharing some functionality with SDT (beyond the use
of breakpoints), the functions dtrace_sdt_peek, dtrace_sdt_poke,
dtrace_sdt_start, and dtrace_sdt_stop are being replaced by generic
versions that can support both SDT and FBT.

In order for FBT probes to be able to function, a change is made in the
trap handling code for breakpoints.  If the ESR code indicates that the
breakpoint is a DTrace FBT breakpoint, we emulate a "mov x29, sp"
instruction because that is the one that got replaced by the breakpoint.
This change does not affect anything beyond FBT related breakpoints.

Orabug: 28643859
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Tom Saeger <tom.saeger@oracle.com>