dtrace: FBT return probe support for arm64
Implementation of the FBT return probes for arm64.  They work similar
to FBT entry probes, i.e. by means of a breakpoint that is placed on
'ret' instructions in functions that can be probed for entry.  It also
provides emulation for the 'ret' instruction right before resuming
execution after the breakpoint (used for FBT only).

Two additional functions are added to the FBT blacklist for arm64
because the architecture provides a specific clocksource that will
be queried during probe processing.

Orabug: 28826056
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Tom Saeger <tom.saeger@oracle.com>