dtrace: exclude init and exit sections for sdtinfo
DTrace SDT probing is not currently supported for module init and exit
sections.  Due to a conflict between kmodsdt and dtrace_sdt.sh processing
for SDT probe information, it could happen that an external reference to
a local function slipped into the sdtinfo data structures, causing a
module load problem because the symbol was deemed unresolvable.

We now correctly skip init and exit text sections for modules.

Orabug: 28808083
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>