ctf: generate CTF information for the kernel from DWARF
This introduces a new tool, dwarf2ctf, which runs whenever 'make ctf' is
run, extracting information on the kernel's types and global variables
from the DWARF-format debug information in the kernel build tree,
deduplicating it, and emitting it in Sun's Compact Type Format into a
mmappable type archive named vmlinux.ctfa, which is installed at 'make
install' time into /lib/modules/$(uname -r)/kernel/.  Out-of-tree
modules cannot participate in this mechanism since the file is already
written: CTF information for such modules is instead linked into such
modules at build time as new sections named .SUNW_ctf (as a result, most
of the build-time machinery for this is in scripts/Makefile.modpost).
Care should be taken not to strip such sections into debug RPMs (they
are small enough that this should not be a problem).

(Note: This tool is now deprecated in favour of the CTF generator
in GCC 13+ and the CTF deduplicator built in to GNU ld.  See the
following commits.)

Within the ctfa file, the type information is divided into a shared
repository, containing all types used by more than one module, CTF for
the core kernel, and separate CTF for each module built, whether or not
this module has been compiled in or not: if a file *could* be built as a
module, it will be considered to be a module from the perspective of CTF
file emission (and kallmodsyms: see the next commit).  This ensures that
external consumers such as DTrace always find types for a given module
in the same place, regardless of the local kernel configuration, as long
as that module is present at all, assisting in portability of D scripts
between installations.  objdump --ctf --ctf-parent=shared_ctf from trunk
binutils can be used to inspect ctfa files.

This process needs a pair of new files, objects.builtin (which lists all
object files that are unconditionally built into the kernel and cannot
be built as modules) and modules_thick.builtin (added by the commits
immediately preceding this one), which maps from the thin archives
that make up built-in modules to their constituent object files.
Taken together, these files let dwarf2ctf determine whether a given
object file linked into vmlinux.o is part of a module, and if so,
which one.

There is a single manually-maintained blacklist of structure members
dwarf2ctf cannot handle in scripts/dwarf2ctf/member.blacklist: this is
used to identify structure members which have different definitions in
different object files even though they are defined in the same location
in the same source file, usually due to preprocessor magic.  (Currently,
the only item in this list is present for example purposes only, since
the file in question was recently removed from the kernel: dwarf2ctf can
these days identify most members needing blacklisting automatically, and
will fail with an error if it needs more help.  It is quite possible that
dwarf2ctf will fail on make allyesconfig kernel configurations and other
extreme cases: I hope to track all such bugs down in time.)

This introduces new kernel build-time dependencies on elfutils, zlib,
glib, and the new libdtrace-ctf package (shared with DTrace userspace).
No new runtime dependencies are introduced.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Signed-off-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>