dtrace: modular components and x86 support
This implements the core DTrace module (including the entire DIF
interpreter and support for all built-in D variables and functions) and
one test provider, dt_test.ko.  It uses the machinery added in the last
few commits.  An x86 implementation of the architecture-dependent parts
is also added so that one platform at least can compile it.

At this stage, almost no probes will exist: they are added by the
following commits, that add providers and SDT probes.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Signed-off-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>
Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Signed-off-by: David Mc Lean <david.mclean@oracle.com>
Signed-off-by: Vincent Lim <vincent.lim@oracle.com>