dtrace: USDT and pid providers
For historical reasons, these are provided in a module named
fasttrap.ko.

Much of this is arch-dependent code for jump-table detection and
implementation of globbed pid probes ("probe everything you can in this
function"), as well as arch-dependent code to look up arguments and code
to ensure that dropping a kprobe in a single process does not affect
other processes running from the same binary.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Signed-off-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>
Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Signed-off-by: David Mc Lean <david.mclean@oracle.com>
Signed-off-by: Vincent Lim <vincent.lim@oracle.com>