dtrace: USDT and pid provider core and x86 components
This implements the core (linked-in) machinery needed for userspace
statically-defined tracepoints (for historical reasons, known as
'fasttrap' by DTrace) and for the pid provider, which allows USDT
probes to be dropped at addresses that do not correspond to symbols,
at locations named as a symbol plus an offset.  Both are implemented in
terms of kprobes.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Signed-off-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>
Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Signed-off-by: David Mc Lean <david.mclean@oracle.com>
Signed-off-by: Vincent Lim <vincent.lim@oracle.com>