dtrace: function boundary tracing (FBT) core and x86 components
This commit implements the core components needed for FBT tracing.
Unlike ftrace we allow the tracing of very large numbers of functions at
once: the intent is that the system should still be stable when every
eligible function in the kernel is traced simultaneously.  Functions
that are not safe for this (because e.g. they are used in trap handling,
or by functions called by the DTrace module itself during probe
processing) are (semi-manually) blacklisted from being probed.

As part of this, a treewide change to the prototype of traps is started:
they all return 0 by default now, with a nonzero return value indicating
that the trap happened as a result of an FBT probe: the return value is
the opcode atop which the trap was originally placed for later emulation.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Signed-off-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Signed-off-by: David Mc Lean <david.mclean@oracle.com>
Signed-off-by: Vincent Lim <vincent.lim@oracle.com>