add DTRACEACT_PCAP to capture packet data for later pcap_dump()
to support libpcap-based packet capture in DTrace, we add a DTRACEACT_PCAP
action which records capture time, full packet length and as much of the
packet as the buffer allows. The linear portion of the skb is copied,
while the non-linear skbs are collected by walking skb page fragments.

Changes since v1: skb fragment processing was not taking the possibility
of compound pages into account.  Importing a slightly-modified version of
the skb_frag_foreach_page() macro resolves the issues and skb fragments
can be read successfully also.  Also moved the pcap functionality into an
inline function to avoid over-indentation.

Orabug: 28951771

Signed-off-by: Alan Maguire <alan.maguire@oracle.com>