fix ip provider inconsistencies between IPv4/IPv6
There are a number of inconsistencies in how IPv4 and IPv6 drop probes fire
(or do not), and it would be good to simplify the code where possible to
minimize the number of drop points.

Changes since v1: as per Darren's review, ensured "dropreason" is a
const char *, and avoided preassigning it for clarity.

Orabug: 28956807

Signed-off-by: Alan Maguire <alan.maguire@oracle.com>
Reviewed-by: Darren Kenny <darren.kenny@oracle.com>