ctf, kallsyms: prevent duplication in modules_thick.builtin
We were duplicating object file names in modules_thick.builtin because
*-objs can sometimes contain names that are already in *-y and *-Y.

This doesn't always happen, and rather than compensating for the precise
cases when it happens (introducing the possibility of breakage when that
changes), just sort the list to eliminate duplicates.  We don't care
about the ordering of object file names anyway.

Reported-by: Eugene Loh <eugene.loh@oracle.com>
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: David  Mc Lean <david.mclean@oracle.com>