ctf: potential refactoring (may be worse than what it replaces)
Split out type DIE computation from type_id_internal() into another
function. This is barely any more readable, but it *does* reduce
type_id_internal() to nothing but a 'decorate with location and type ID
of this DIE in particular' function -- splitting out everything to do
with the parent type DIE, which is a bit squirrelly because of the
requirement to pass bit_size and bit_offset overrides down.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: David  Mc Lean <david.mclean@oracle.com>