dtrace: de-typedef
Done largely via new scripts in scripts/coccinelle/dtrace, one to remove
struct typedefs, the other to remove enum typedefs.

(Manual adjustments atop this: consistently reflow all
structure-returning function definitions to put the return type on the
previous line, repair tabbing of declarators, and general adjustments to
fit within 80 chars.)

This is unadorned, unmodified Coccinelle output. Its indentation is
frequently screwed up and it breaks the DTrace uapi: DTrace userspace
will not compile against this.  This is fixed up in subsequent commits.

Applied via

rm /tmp/detypedef-arm.diff

for scr in typedef-elision enum-elision do

    spatch --cocci-file scripts/coccinelle/dtrace/$src.cocci \
	   --defined 'IS_ENABLED(CONFIG_DT_FASTTRAP)' --prevent-kr \
	   --patch . -I ./arch/x86/include -I ./arch/x86/include/asm \
	   -I ./arch/x86/include/generated -I ./include \
	   -I ./arch/x86/include/uapi -I ./arch/x86/include/generated/uapi \
	   -I ./include/uapi -I ./include/generated/uapi \
	   --include ./include/linux/kconfig.h include/linux/dtrace*.h \
	   include/dtrace/*.h include/uapi/linux/dtrace/*.h \
	   arch/x86/dtrace/include/dtrace/*.h dtrace/*.h \
	   dtrace/*.c kernel/dtrace/*.c arch/x86/dtrace/*.c \
	   arch/x86/kernel/dtrace*c kernel/sched/sched.h | git apply

    spatch --cocci-file scripts/coccinelle/dtrace/$src.cocci \
	   --defined 'IS_ENABLED(CONFIG_DT_FASTTRAP)' --prevent-kr \
	   --patch . -I ./arch/arm64/include -I ./arch/arm64/include/asm \
	   -I ./arch/arm64/include/generated -I ./include \
	   -I ./arch/arm64/include/uapi -I ./arch/arm64/include/generated/uapi \
	   -I ./include/uapi -I ./include/generated/uapi \
	   --include ./include/linux/kconfig.h include/linux/dtrace*.h \
	   include/dtrace/*.h include/uapi/linux/dtrace/*.h \
	   arch/arm64/dtrace/include/dtrace/*.h dtrace/*.h \
	   arch/arm64/dtrace/*.c arch/arm64/kernel/dtrace*c \
	   kernel/sched/sched.h >> /tmp/detypedef-arm.diff
done

    # edit /tmp/detypedef.arm to remove all changes outside arch/arm64,
    # then git apply

The weird editing in the latter is because we have to rerun it over the
headers, since that's where most of the typedefs are picked up from.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>