dtrace: manual removal of typedefs missed by Coccinelle
Some typedefs had forms, or were in locations, that meant that the
Coccinelle scripts to do the detypedeffing missed them.  Also,
some typedefs survive despite the erasure of all their uses (mostly
the ones called out by hand in the Coccinelle scripts).

Fix them up by hand (and in places with one-liner Coccinelle scripts).

Fixed up:

  dtrace_module_t
  fasttrap_id_t
  fasttrap_probe_t
  fasttrap_tracepoint_t
  cred_t
  cyc_handler
  cyc_time
  cyc_omni_handler
  cyclic_t
  dt_fbt_bl_entry
  sdt_probedesc

Typedefs with no remaining uses elided:
  dtrace_ecb_t
  dtrace_state_t
  dtrace_task_t

In places, a little light reformatting is done too (see later commits in
this series).  This often makes the formatting worse: the remaining
commits fix it up again.  (It was that, or merge the two commits, and
that seemed a worse choice.)

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>