dtrace: fix catastrophic Coccinelle-induced mistabbing
The DTrace style (inherited from Sun) is to tab out declarator blocks to
a consistent width.  This is somewhat disastrous when combined with very
long type names.  Comments are often also mispositioned, and some
definitions of functions both taking and returning structures are now
very long.  Coccinelle also messed up line-wrapping here and there,
introducing staircasing.

Fix up, retabbing, detabbing, and moving initializers into
non-initializer code as needed.  The only priority here is to keep
identifier names roughly columnar: if one long identifier sticks out of
a column, or a decl block is split in two to allow a column to shrink in
width dramatically (while remaining columnar), do so, as long as it
improves readability (which it usually does, by bringing variable names
and their types closer together).  Equally, if a lot of identifiers are
the same width and can be made columnar using only a space or two,
prefer that to introducing tabs and yanking things further apart.
No priority is given to retaining pre-detypedef decl tabbing for its own
sake: some of it is very easy to make more readable, so do so.
(Long    runs of space   without need
 make text quite   hard to read.)

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>