ctf: speed up make ctf by 40--50% by eschewing unnecessary ctf_update()s
libdtrace-ctf master, to become 1.2.0, makes ctf_update() less necessary
than it was: you only need to call it when doing things other than a run
of ctf_add_*() calls.  No longer do you need to call it when type
additions return -ECTF_BADID: they shouldn't do so at all unless you
pass in an ID that really *is* bad (something older libdtrace-ctf wasn't
validating very well, but it is now).  libdtrace-ctf master sets a
define to allow callers to observe when this feature is available, so
we can continue to work with older libdtrace-ctf at the same old speed.

In conjunction with libdtrace-ctf master , this speeds up 'make ctf' by
on the order of 40--50%.

(Also, shuffle some nearby code around a bit to avoid accessing a freed
pointer when intern(id) is called with an already-interned atom.)

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>