dtrace: report signals discarded because of process death
As of commit 35634ffa1751b6efd8cf75010b509dcb0263e29b, signals sent to
processes that are already dying are silently discarded, skipping over
DTrace probe firing.

Avoid this by arranging to fire a signal-send with NULL args[2] and
args[3] (the siginfo and handler parameters).  This combination is
unique to this situation, so can be used to detect that a received
signal was not really *sent* anywhere (but was also not exactly
discarded in the ignored-signal signal-discard sense) because the
process was dying.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>