ctf: adjust for lack of top-level vmlinux.o
As of commit dee9495303f2c6d63a20c43a26420765909898eb, the top-level
thin archive built-in.a is removed, as is vmlinux.o.  'make ctf' was
relying on this to get built-in objects from: without it we end up with
a vmlinux.ctfa without a vmlinux.ctf in it, covering only modules, not
the core non-modular kernel.  Thankfully we already have a list of
built-in real object files computed from KBUILD_VMLINUX_OBJS, and
already determine which object file something is in using the DWARF, not
filenames, so it is easy to adjust: simply pass in the contents of
objects.builtin instead of vmlinux.o or a huge list of object files, and
we end up with a properly-formed vmlinux.ctfa again.

In another irregularity, the evolution of the code ended up in a state
where we were expanding thin archives into their components if 'make
ctf' was invoked one way and using objects.builtin (which already did
this at the top level) otherwise.  Standardize on one approach:
objects.builtin plus all the modules.

While we're at it, reformulate the filelist creation to speed it up:
running file(1) several thousand times takes about 20% of the total
time taken by 'make ctf', which is not cheap itself.  We can rely on
*.builtin being textual that needs concatenation, *.a being readable
by 'ar t' (whether or not they are thin archives), and everything else
being an object file, so there's no need to run file(1) to prove that.
(Double up on the builtins just in case the top-level builtin file
ever gets a name not ending in *.builtin.)

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>