dtrace: fix tcp:::send args[2]
The tcp:::send probe provides the wrong args[2] on the kernel end.
In net/ipv4/tcp_output.c:

        DTRACE_TCP(send,
                   struct sk_buff * :  pktinfo_t *, skb,
                   struct sock * : csinfo_t *, sk,
                   __dtrace_tcp_void_ip_t * : ipinfo_t *, ip_hdr(skb),
                   struct tcp_sock * : tcpsinfo_t *, tp,
                   struct tcphdr * : tcpinfo_t *, tcp_hdr(skb),
                   int : tcplsinfo_t *, sk->sk_state, int, sk->sk_state,
                   int, DTRACE_NET_PROBE_OUTBOUND);

That "ip_hdr(skb)" is known to be garbage, and userspace expects this
to be indicated with a NULL value.  This problem has avoided detection
so long because if the version (4 or 6) extracted from this garbage is
neither 4 nor 6, a reliable fallback strategy is used.

Change the "ip_hdr(skb)" to "NULL".

Orabug: 29594231
Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Alan Maguire <alan.maguire@oracle.com>