ctf: kernel build with -gctf for CTF generation using GCC (sans ld atm)
Add two new CONFIGs for a CTF enabled toolchain and a kernel build with
-gctf or -gt and -lctf respectively:
  - new config HAVE_CTF_TOOLCHAIN which is set if the toolchain supports CTF
    generation.
  - new config CTF_USING_BINUTILS to control whether CTF generation is to be
    done using the GNU Toolchain.  This config is unset by default.

...
CONFIG_HAVE_CTF_TOOLCHAIN=y
CONFIG_CTF=y
\# CONFIG_CTF_USING_BINUTILS is not set
...

Also, -gctf is added to the compilation flags if the toolchain is CTF
capable: the old, non-upstreamed flag, -gt, is also detected.
The uptake of -lctf is not included in this patch.

Signed-off-by: Indu Bhagat <indu.bhagat@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>