ctf: discard CTF from the vDSO
The vDSO is meant to be no more than one page long: but CTF is getting
left in it, making it much longer than that on all arches that include C
code in the vDSO (which is most of them).  Amazingly, this doesn't cause
instant disaster or indeed many visible problems at all: but it does
cause GDB to infloop on opening anything and it is definitely both a
mistake and a pointless waste of space, since the vDSO cannot contribute
types to any running program.

Exclude it on all appropriate arches.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Victor Erminpour <victor.erminpour@oracle.com>