ctf: adjust to upcoming binutils ctf_link_add_ctf API change
The unreviewed intermediate state of the binutils CTF deduplicator which
got into some Linux distributions (for example, older Oracle Linux 8)
broke the API for ctf_link_add_file.  This breakage is gone in the
latest release, so we have to adapt without breaking if the old binutils
is used so we don't have to upgrade everything in sync.

Thankfully the unreviewed intermediate backport also contains #defines
which do not exist in the latest version: one such is CTF_LINK_FINAL.
So we can use this to pick the right API.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Jack Vogel <jack.vogel@oracle.com>