dtrace: fix buffer overflow in profile provider
This strcpys the probe name into place without ever checking its length,
and accepts names much longer than the space allocated.

Fix by using DTRACE_NAMELEN (since this is the length of a probe name,
and the name in question is the name of a probe), checking the length,
and failing hard if it's too large.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>