pidfd: make ptrace and pidfd work together
These changes are a proof-of-concept to show how pidfd might be usable
to determine when a waitpid on a ptraced process would yield a response
without blocking.  This can be useful if you want a 'ptrace helper'
thread (or process) that accepts messages regarding what ptrace work
needs doing and also ptrace-monitors other processes to do that work.
It is obviously best to do the interprocess messaging part via a file
descriptor, but that means you have to poll() it, and if you're stuck in
poll() you can't wake up in time when one of your ptrace-monitored
processes needs attention, since waitpid does not yield an fd.  This
situation is tailor-made for pidfds!

Up until now I've handled this with a really rather ugly hack on the old
(long-rejected) waitfd, see
e.g. https://github.com/oracle/dtrace-linux-kernel/commit/81e92377b8e2f89fc0cde898738321dd869f21d2.
This commit is nasty in all sorts of ways and pidfds would be better in
every respect were it not that right now pidfds don't wake up pollers in
response to ptrace-induced waits.[1]

This patch fixes it, without breaking pidfd's contract: we lift the
restriction that pidfds can only be used on thread-group leaders iff a
new PIDFD_THREAD flag is passed to pidfd_open() (necessary since ptrace
wakes are thread-directed).

Of necessity, since we save this flag nowhere (it's only used to affect
validation at pidfd_open() time) we also cause poll()s of pidfds to wake
up whenever any thread exits; since spurious polls can happen for all
sorts of reasons (other threads using the same fds or dupped fds, other
processes doing the same, other threads waiting on a child before the
poller can, etc) this is presumably harmless.  (If it's not, I'm happy
to adjust things so that such wakeups only happen when PIDFD_THREAD is
set, if you have any idea how to do that.)

We do reduce the frequency of such things by using the old approach
whenever this pid is the pid of a thread group leader and the process is
not being ptraced.  This is a bit ugly and may be a step too far: it
more or less prevents you using PIDFD_THREAD for anything but ptracers
for the sake of back-compatibility (so maybe we should call PIDFD_THREAD
PIDFD_PTRACE instead?).  But it does prevent a change in poll behaviour
on non-PIDFD_THREAD pidfds except in the rare case of your ptracing a
thread group leader and someone *else* ptracing it (and even then you
only get spurious wakeups).

A trivial testcase (with a trivial singlethreaded child, and no
messaging layer) is provided so you can see the sort of thing I'm trying
to make work.

[1] It is also true that pidfds integrate with *waitid()*, not
    waitpid(), and waitid() cannot be used to interact with ptrace --
    but that's OK, because you can use a pidfd awakening from poll as
    proof that waitpid() would return with ptraced info if called, and
    then call waitpid() on the PID directly, without going via the
    pidfd.  This is not ideal, since it would be nice to be able to use
    waitid with P_PIDFD to get the info directly without also needing
    the associated process ID, but as long as waitid() masks out and
    throws away information returned by waitpid() for ptraced waits, it
    is the best we can do.  This is implied by the ptrace manpage, which
    never mentions waitid, because only waitpid works.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>