kbuild: bring back tristate.conf
tristate.conf was dropped because it is not needed to build
modules.builtin any more, and doing so avoids one round of recursion
through the build tree to build it.  But it has one property that can be
obtained in no other way in the current tree: it provides a
machine-readable record of whether a module is tristate or not.
(modules.builtin.objs, just added, uses modinfo, which is recorded in
the source files themselves, but it is Kconfig that actually controls
whether something can be built as a module.)

So bring it back for this purpose.  (Thanks to the refactoring in
the 5.16 timeframe, this is basically a reimplementation of commit
8b41fc4454e36fbfdbb23f940d023d4dece2de29 rather than a simple
reversion.)

A verifier that uses it will be added in the next commit.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Victor Erminpour <victor.erminpour@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>