kbuild: generate an address ranges map at vmlinux link time
This emits a new file, .tmp_vmlinux.ranges, which maps address
range/size pairs in vmlinux to the object files which make them up,
e.g., in part:

0x0000000000000000 0x30 arch/x86/kernel/cpu/common.o
0x0000000000001000 0x1000 arch/x86/events/intel/ds.o
0x0000000000002000 0x4000 arch/x86/kernel/irq_64.o
0x0000000000006000 0x5000 arch/x86/kernel/process.o
0x000000000000b000 0x1000 arch/x86/kernel/cpu/common.o
0x000000000000c000 0x5000 arch/x86/mm/cpu_entry_area.o
0x0000000000011000 0x10 arch/x86/kernel/espfix_64.o
0x0000000000011010 0x2 arch/x86/kernel/cpu/common.o
[...]

This will be used by kallmodsyms to let us associate symbols with
built-in modules at address-range granularity (i.e., space-efficiently),
and to let us disambiguate symbols with identical name and module by
annotating them with the translation unit they come from.

In my simple tests this seems to work with clang too, but if I'm not
sure how stable the format of clang's linker mapfiles is: if it turns
out not to work in some versions, the mapfile-massaging awk script added
here might need some adjustment.

CONFIG_LTO_CLANG by default optimizes by working from the LTOed
vmlinux.o file: but this doesn't work for kallmodsyms, as the resulting
mapfile lists vmlinux.o as the source object file in all cases, ruining
its attempt to disambiguate symbols using object file names.  So we
suppress that optimization in this case.

We also suppress it when IBT is enabled, but IBT *requires* use of the
intermediate vmlinux.o, since that's the .o that objtool has run over.
We lift this restriction in the next commit.

(There are similar problems with everything else that uses ld -r: since
this amounts, in total, to a few parts of KVM on aarch64, I haven't
implemented a general ld -r fix: the fix in the next commit is specific
to vmlinux.o.)

Orabug: 29891866
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>