ctf: generate CTF information for the kernel
This introduces a new config option, CONFIG_CTF, which causes the
compiler to be run with -gctf to emit information on the kernel's types
and global variables, and a tool, ctfarchive, which runs whenever 'make
ctf' is run, deduplicating that information into a mmappable type archive
named vmlinux.ctfa, which is installed at 'make install' time into
/lib/modules/$(uname -r)/kernel/.

Out-of-tree modules cannot participate in this mechanism since the file
is already written: the compiler emits .ctf sections into the modules
themselves, and the linker deduplicates them.  ctfarchive uses the same
code (in libctf) to do the deduplication into vmlinux.ctfa as the linker
does for out-of-tree modules: there is no code duplication.  strip does
not strip such sections out, but additional care should be taken not to
strip such sections into debug RPMs (they are small enough that this
should not be a problem).

Within the ctfa file, the type information is divided into a shared
repository, containing all types used by more than one module, CTF for
the core kernel, and separate CTF for each module built, whether or not
this module has been compiled in or not: if a file *could* be built as a
module, it will be considered to be a module from the perspective of CTF
file emission (and kallmodsyms).  This ensures that external consumers
such as DTrace always find types for a given module in the same place,
regardless of the local kernel configuration, as long as that module is
present at all, assisting in portability of D scripts between
installations.  objdump --ctf --ctf-parent=shared_ctf from trunk
binutils can be used to inspect ctfa files.

This process needs a pair of new files, objects.builtin (which lists all
object files that are unconditionally built into the kernel and cannot
be built as modules) and modules_thick.builtin (added by the commits
immediately preceding this one), which maps from the thin archives that
make up built-in modules to their constituent object files.  Taken
together, these files let ctfarchive and the ctf_link machinery
determine whether a given object file linked into vmlinux.o is part of a
module, and if so, which one.

This introduces new kernel build-time dependencies on a sufficiently new
binutils, a CTF-capable GCC, and libctf: if any of these is not found,
CONFIG_CTF is disabled.

Orabug: 29891866
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Signed-off-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>
Reviewed-by: Victor Erminpour <victor.erminpour@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>